requirements.txt
pytest~=7.2.0
requests~=2.28.1
setuptools~=65.5.1
python-dotenv~=0.21.0
environs~=9.5.0
jira~=3.4.1
singleton-decorator~=1.0.0
allure-pytest~=2.11.1
